started to work on datacollector

+++ visualswarm/vision/vprocess.py
@@ -191,7 +191,7 @@ def high_level_vision(raw_vision_stream, high_level_vision_stream, visualization
 
                 frame_rgb = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
                 frame_resized = cv2.resize(frame_rgb, (width, height))
-                input_data = np.expand_dims(frame_resized, axis=0)
+                input_data = np.expand_dims(frame_resized, axis=0).astype("float32")
 
                 logger.info(f"dim: {input_data.shape}, min: {np.min(input_data)}, max: {np.max(input_data)}")
